extract applyTemplate into own class
